+++ Muse/Model/Order.cs
@@ -15,7 +15,7 @@ namespace Muse.Model {
         [Required]
         public int Quantity { get; set; }
 
-        public int Total { get { return Product.Price * Quantity; } }
+        public decimal Total { get { return Product.Price * Quantity; } }
 
         [Required]
         public DateTime CreatedAt { get; set; }
