+++ Muse/ViewModel/OrderViewModel.cs
@@ -10,10 +10,10 @@ namespace Muse.ViewModel {
 
         public string ProductName { get; set; }
 
-        public int ProductPrice { get; set; }
+        public decimal ProductPrice { get; set; }
 
         public int Quantity { get; set; }
  
-        public int Total { get; set; }
+        public decimal Total { get; set; }
     }
 }
